+++ modules/canvaskit/npm_build/types/canvaskit-wasm-tests.ts
@@ -440,6 +440,8 @@ function paintTests(CK: CanvasKit, colorFilter?: ColorFilter, imageFilter?: Imag
     paint.setImageFilter(imageFilter);
     paint.setMaskFilter(maskFilter);
     paint.setPathEffect(pathEffect);
+    // @ts-expect-error
+    paint.setShader(colorFilter);
     paint.setShader(shader);
     paint.setStrokeCap(CK.StrokeCap.Round);
     paint.setStrokeJoin(CK.StrokeJoin.Miter);